requirements.txt
selenium==4.20.0
selene==2.0.0rc4
pytest==7.4.0
allure-pytest==2.13.2
requests==2.32.3
python-dotenv